requirements.txt
jinja2
yagmail
requests
python-dotenv
feedparser
beautifulsoup4
aiohttp
openai
